requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
babel==2.17.0
beautifulsoup4==4.13.4
billiard==4.2.1
bleach==6.2.0
cachetools==6.1.0
celery==5.5.3
certifi==2025.8.3
cffi==1.17.1
channels==4.3.1
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
constantly==23.10.4
coverage==7.10.3
crispy-bootstrap5==2025.6
cryptography==45.0.6
daphne==4.2.1
decorator==5.2.1
diff-match-patch==20241021
distro==1.9.0
Django==5.2.5
django-allauth==65.10.0
django-appconf==1.1.0
django-bootstrap5==25.2
django-crispy-forms==2.4
django-debug-toolbar==6.0.0
django-environ==0.12.0
django-filter==25.1
django-formtools==2.5.1
django-guardian==3.0.3
django-import-export==4.3.9
django-money==3.5.4
django-redis==6.0.0
django-unfold==0.64.2
django-unicorn==0.62.0
djangorestframework==3.16.1
fido2==2.0.0
gitdb==4.0.12
GitPython==3.1.41
groq==0.31.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
iniconfig==2.1.0
kombu==5.5.4
Markdown==3.8.2
martor==1.7.11
oauthlib==3.3.1
orjson==3.11.2
packaging==25.0
pluggy==1.6.0
prompt_toolkit==3.0.51
psycopg==3.2.9
psycopg-binary==3.2.9
py-moneyed==3.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pyOpenSSL==25.1.0
pytest==8.4.1
pytest-cov==6.2.1
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
qrcode==8.2
redis==5.2.1
requests==2.32.4
sentry-sdk==2.35.0
service-identity==24.2.0
setuptools==75.6.0
shortuuid==1.0.13
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
sqlparse==0.5.3
tablib==3.8.0
Twisted==25.5.0
txaio==25.6.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
whitenoise==6.9.0
zipp==3.23.0
zope.interface==7.2
